Refactor for reintegration of projects.
