+++ src/Demo/NetEx.WinForms.ProgressDialogDemo/Demo.ico
