+++ src/Demo/NetEx.WinForms.ProgressDialogDemo/ProgressDialog/ProgressDialogForm.resx
