+++ src/NetEx.sln
@@ -3,9 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34728.123
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetEx.WinForms.ProgressDialog", "NetEx.WinForms.ProgressDialog\NetEx.WinForms.ProgressDialog.csproj", "{DAD298FB-BB35-417C-B108-65D377ACDAEB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Windows.Forms.ProgressDialog", "System.Windows.Forms.ProgressDialog\System.Windows.Forms.ProgressDialog.csproj", "{DAD298FB-BB35-417C-B108-65D377ACDAEB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetEx.WinForms.ProgressDialogDemo", "NetEx.WinForms.ProgressDialogDemo\NetEx.WinForms.ProgressDialogDemo.csproj", "{91937276-F353-4B0A-BBA8-92FE3AD58E6D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetExDemo", "Demo\NetEx.WinForms.ProgressDialogDemo\NetExDemo.csproj", "{91937276-F353-4B0A-BBA8-92FE3AD58E6D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demo", "Demo", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -24,6 +26,9 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{91937276-F353-4B0A-BBA8-92FE3AD58E6D} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4C5D94DD-DB92-4B78-8E45-AD7C6DA0516A}
