+++ src/System.Windows.Forms.ProgressDialog/AnimationResource.cs
