+++ src/System.Windows.Forms.ProgressDialog/Internal/IProgressDialog.cs
