+++ src/System.Windows.Forms.ProgressDialog/Internal/PROGDLG.cs
