+++ src/System.Windows.Forms.ProgressDialog/Internal/WindowsProgressDialog.cs
