+++ src/System.Windows.Forms.ProgressDialog/NativeMethods.cs
