+++ src/System.Windows.Forms.ProgressDialog/ProgressDialog.cs
