+++ src/System.Windows.Forms.ProgressDialog/ProgressDialogProgressBarStyle.cs
