+++ src/System.Windows.Forms.ProgressDialog/ToolboxBitmap.png
