+++ src/System.Windows.Forms.ProgressDialog/icon.png
